'verify.DiffStringMaps' -> 'flex.DiffStringMaps'.
