+++ internal/flex/flex.go
@@ -319,3 +319,30 @@ func (s Set[T]) Difference(ns Set[T]) Set[T] {
 	}
 	return result
 }
+
+// DiffStringMaps returns the set of keys and values that must be created, the set of keys
+// and values that must be destroyed, and the set of keys and values that are unchanged.
+func DiffStringMaps(oldMap, newMap map[string]interface{}) (map[string]*string, map[string]*string, map[string]*string) {
+	// First, we're creating everything we have
+	add := map[string]*string{}
+	for k, v := range newMap {
+		add[k] = aws.String(v.(string))
+	}
+
+	// Build the maps of what to remove and what is unchanged
+	remove := map[string]*string{}
+	unchanged := map[string]*string{}
+	for k, v := range oldMap {
+		old, ok := add[k]
+		if !ok || aws.StringValue(old) != v.(string) {
+			// Delete it!
+			remove[k] = aws.String(v.(string))
+		} else if ok {
+			unchanged[k] = aws.String(v.(string))
+			// already present so remove from new
+			delete(add, k)
+		}
+	}
+
+	return add, remove, unchanged
+}
