+++ internal/verify/diff_test.go
@@ -4,11 +4,8 @@
 package verify
 
 import (
-	"reflect"
 	"testing"
 	"time"
-
-	"github.com/hashicorp/terraform-provider-aws/internal/flex"
 )
 
 func TestSuppressEquivalentRoundedTime(t *testing.T) {
@@ -60,105 +57,6 @@ func TestSuppressEquivalentRoundedTime(t *testing.T) {
 
 		if !tc.equivalent && value {
 			t.Fatalf("expected test case %d to not be equivalent", i)
-		}
-	}
-}
-
-func TestDiffStringMaps(t *testing.T) {
-	t.Parallel()
-
-	cases := []struct {
-		Old, New                  map[string]interface{}
-		Create, Remove, Unchanged map[string]interface{}
-	}{
-		// Add
-		{
-			Old: map[string]interface{}{
-				"foo": "bar",
-			},
-			New: map[string]interface{}{
-				"foo": "bar",
-				"bar": "baz",
-			},
-			Create: map[string]interface{}{
-				"bar": "baz",
-			},
-			Remove: map[string]interface{}{},
-			Unchanged: map[string]interface{}{
-				"foo": "bar",
-			},
-		},
-
-		// Modify
-		{
-			Old: map[string]interface{}{
-				"foo": "bar",
-			},
-			New: map[string]interface{}{
-				"foo": "baz",
-			},
-			Create: map[string]interface{}{
-				"foo": "baz",
-			},
-			Remove: map[string]interface{}{
-				"foo": "bar",
-			},
-			Unchanged: map[string]interface{}{},
-		},
-
-		// Overlap
-		{
-			Old: map[string]interface{}{
-				"foo":   "bar",
-				"hello": "world",
-			},
-			New: map[string]interface{}{
-				"foo":   "baz",
-				"hello": "world",
-			},
-			Create: map[string]interface{}{
-				"foo": "baz",
-			},
-			Remove: map[string]interface{}{
-				"foo": "bar",
-			},
-			Unchanged: map[string]interface{}{
-				"hello": "world",
-			},
-		},
-
-		// Remove
-		{
-			Old: map[string]interface{}{
-				"foo": "bar",
-				"bar": "baz",
-			},
-			New: map[string]interface{}{
-				"foo": "bar",
-			},
-			Create: map[string]interface{}{},
-			Remove: map[string]interface{}{
-				"bar": "baz",
-			},
-			Unchanged: map[string]interface{}{
-				"foo": "bar",
-			},
-		},
-	}
-
-	for i, tc := range cases {
-		c, r, u := DiffStringMaps(tc.Old, tc.New)
-		cm := flex.FlattenStringMap(c)
-		rm := flex.FlattenStringMap(r)
-		um := flex.FlattenStringMap(u)
-		if !reflect.DeepEqual(cm, tc.Create) {
-			t.Fatalf("%d: bad create: %#v", i, cm)
-		}
-		if !reflect.DeepEqual(rm, tc.Remove) {
-			t.Fatalf("%d: bad remove: %#v", i, rm)
-		}
-		if !reflect.DeepEqual(um, tc.Unchanged) {
-			t.Fatalf("%d: bad unchanged: %#v", i, rm)
 		}
 	}
 }
